moved investigationRule component into dynamicSampling dir
