+++ static/app/components/dynamicSampling/investigationRule.tsx
